Answer randomization and files refactor
Random order for answers display was implemented. Some files was moved into another folders
