+++ model/questionQuery.js
@@ -24,12 +24,12 @@ async function getQuestionObject(index, category) {
 }
 
 // get a posible answer for a question
-async function getChoice(index_question, index_choice, index_round) {
+async function getChoice(index_question, index_round) {
   let question = await getQuestionObject(index_question, index_round);
   let choices = await question.answers;
-  let choice_object = await choices[index_choice];
-  let choice = await choice_object.answer;
-  return choice;
+  // let choice_object = await choices[index_choice];
+  // let choice = await choice_object.answer;
+  return choices;
 }
 
 //Get the question
